Don't include HTML in meta descriptions.
This modifies the way we process the descriptions so that
they're maintained in their text form until they're rendered.

This allows us to decide whether we'd like to materialize them
as HTML (if we're rendering them on the page) or if we'd like
to try to find a text based representation (if they're being
displayed in a `<meta />` tag).